Refactor js into page.js
